Refactored the AssetService testcase
