+++ dao/src/it/java/com/hashmapinc/server/dao/service/sql/AssetServiceSqlTest.java
@@ -16,45 +16,10 @@
  */
 package com.hashmapinc.server.dao.service.sql;
 
-import com.datastax.driver.core.utils.UUIDs;
-import com.hashmapinc.server.common.data.asset.Asset;
-import com.hashmapinc.server.common.data.datamodel.DataModel;
-import com.hashmapinc.server.common.data.id.DataModelObjectId;
 import com.hashmapinc.server.dao.service.BaseAssetServiceTest;
 import com.hashmapinc.server.dao.service.DaoSqlTest;
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.util.List;
 
 @DaoSqlTest
 public class AssetServiceSqlTest extends BaseAssetServiceTest {
-    @Test
-    public void testFindAssetsByDataModelObjectId() {
-        Asset asset = new Asset();
-        asset.setTenantId(tenantId);
-        DataModelObjectId dataModelObjectId = new DataModelObjectId(UUIDs.timeBased());
-        asset.setDataModelObjectId(dataModelObjectId);
-        asset.setName("My asset");
-        asset.setType("default");
-
-        Asset asset2 = new Asset();
-        asset2.setTenantId(tenantId);
-        asset2.setDataModelObjectId(dataModelObjectId);
-        asset2.setName("My asset2");
-        asset2.setType("default");
-
-        Asset savedAsset = assetService.saveAsset(asset);
-        Asset savedAsset2 = assetService.saveAsset(asset2);
-
-        List<Asset> foundAsset = assetService.findAssetsByDataModelObjectId(savedAsset.getDataModelObjectId());
-
-        Assert.assertNotNull(foundAsset);
-        Assert.assertEquals(2, foundAsset.size());
-
-        assetService.deleteAsset(savedAsset.getId());
-        assetService.deleteAsset(savedAsset2.getId());
-    }
-
 
 }
